requirements.txt
kopf
boto3
kubernetes
pyyaml
